requirements.txt
discord.py>=2.0.0
python-dotenv>=0.19.0
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=0.24.0
Pillow>=8.0.0